move AppLogo into Settings
