+++ source/views/settings/components/logo.js
@@ -4,7 +4,7 @@ import * as React from 'react'
 import * as Icons from '@hawkrives/react-native-alternate-icons'
 import glamorous from 'glamorous-native'
 
-import {lookup as getAppIcon} from '../../images/icons/index'
+import {lookup as getAppIcon} from '../../../../images/icons/index'
 
 const LogoImage = glamorous.image({
 	width: 100,
